Added the ability to have more then a single split
This is still mostly hardcoded but it is now possible to define
multiple virtual monitors. Instead of just using the unused bits
as flags, this patch changes it to use them to track each virtual
screen.

This also causes the virtual monitors to be named with a numerical
value instead of a hardcoded _, and also corrects the x position
of any additional screens.